+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { LoginComponent } from './login/login.component';
+import { LoginComponent } from 'app/components/account/login/login.component';
 import { TopNavComponent } from 'app/components/shared/top-nav/top-nav.component';
 import { HomeComponent } from 'app/components/general/home/home.component';
 import { RegisterComponent } from 'app/components/account/register/register.component';
@@ -13,7 +13,9 @@ import { AboutUsComponent } from './components/general/about-us/about-us.compone
 import { PageNotFoundComponent } from './components/general/page-not-found/page-not-found.component';
 
 const routes: Routes = [
-  { path: 'home', component: HomeComponent },
+  { path: 'home', component: HomeComponent, children: [
+    {path: 'nav', component: TopNavComponent}
+  ] },
   { path: 'register', component: RegisterComponent },
   { path: 'login', component: LoginComponent },
   // {
