+++ src/app/components/account/login/login.component.html
@@ -1,3 +1,5 @@
-<p>
-  login works!
-</p>
+<div [hidden]="isSignedIn()"
+     id="fbui-auth-container"></div>
+<pre *ngIf="data">
+  {{ data | json }}
+</pre>
