+++ src/app/components/account/login/login.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthService } from '../../../services/auth.service';
+import {UserInfoOpen} from '../../../shared/class/user-info';
 
 @Component({
   selector: 'cos-login',
@@ -6,8 +8,21 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./login.component.scss']
 })
 export class LoginComponent implements OnInit {
+  title = 'cos';
+  data: {};
+  constructor(private AuthSvc: AuthService) {
+    this.AuthSvc.authInfo.subscribe(data => {
+      this.data = data;
+    });
+    this.AuthSvc.startUi('fbui-auth-container');
 
-  constructor() { }
+
+  }
+
+  isSignedIn() {
+    const signedIn = this.AuthSvc.isSignedIn();
+    return signedIn;
+  }
 
   ngOnInit() {
   }
