--- src/app/login/login.component.html
@@ -1,5 +1,0 @@
-<div [hidden]="isSignedIn()"
-     id="fbui-auth-container"></div>
-<pre *ngIf="data">
-  {{ data | json }}
-</pre>
